finish lab day 2
