+++ xmem_setup/xmem_setup/Debug/makedep.mk
@@ -6,3 +6,5 @@ main.c
 
 USARTdriver.c
 
+xmem.c
+
